tagging version 1.1.8
git-svn-id: https://plugins.svn.wordpress.org/geodirectory/tags/1.1.8@951167 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ geodirectory-functions/post_functions.php
@@ -24,6 +24,7 @@ function geodir_set_postcat_structure($post_id, $taxonomy, $default_cat = '' , $
 		
 	if(isset($category_str) && empty($category_str))
 	{
+		
 		$post_cat_str = '';
 		$post_categories = array();
 		if(isset($post_cat_array) && is_array($post_cat_array) && !empty($post_cat_array)){
@@ -34,6 +35,18 @@ function geodir_set_postcat_structure($post_id, $taxonomy, $default_cat = '' , $
 		$post_categories[$taxonomy] = $post_cat_str;
 		$category_str = $post_categories;
 	}
+	
+	$change_cat_str = $category_str[$taxonomy];
+	
+	$default_pos = strpos($change_cat_str, 'd:');
+	
+	if ($default_pos === false) {
+			
+		$change_cat_str = str_replace($default_cat.',y:', $default_cat.',y,d:', $change_cat_str);
+		
+	}
+	
+	$category_str[$taxonomy] = $change_cat_str;
 	
 	update_post_meta($post_id, 'post_categories', $category_str);		
 	
@@ -634,7 +647,7 @@ function geodir_save_post_images($post_id = 0, $post_image = array(), $dummy = f
 		
 		$geodir_uploadpath = $uploads['path'];
 		$geodir_uploadurl = $uploads['url']; 	
-		$sub_dir = $uploads['subdir'];
+		$sub_dir = isset($uploads['subdir']) ? $uploads['subdir'] : '';
 		
 		$invalid_files = array();
 		$postcurr_images = array();
@@ -728,7 +741,7 @@ function geodir_save_post_images($post_id = 0, $post_image = array(), $dummy = f
 						}elseif(file_exists($uploads['basedir'].$curr_img_dir.$filename)){$uploaded_file = true;$file_path = $curr_img_dir.'/'.$filename;}
 						
 						
-						if($curr_img_dir != $geodir_uploaddir)	
+						if($curr_img_dir != $geodir_uploaddir && file_exists($img_path))	
 							unlink($img_path);
 						
 					}		
@@ -933,9 +946,9 @@ function geodir_get_featured_image( $post_id = '', $size = '' ,$no_image = false
 	 $img_arr = array();
 	 
 	 $file_info = pathinfo($file);
-			
+			$sub_dir = '';
 		if($file_info['dirname'] != '.' && $file_info['dirname'] != '..')
-			$sub_dir = $file_info['dirname'];
+			$sub_dir = stripslashes_deep($file_info['dirname']);
 		
 			$uploads = wp_upload_dir(trim($sub_dir, '/')); // Array of key => value pairs	
 			$uploads_baseurl = $uploads['baseurl'];
@@ -1038,6 +1051,7 @@ function geodir_get_images($post_id = 0, $img_size='', $no_images =false, $add_f
 	global $wpdb;
 	
 	$not_featured = '';
+	$sub_dir = '';
 	if(!$add_featured)
 		$not_featured = " AND is_featured = 0 ";
 	
@@ -1062,9 +1076,8 @@ function geodir_get_images($post_id = 0, $img_size='', $no_images =false, $add_f
 			
 			$file_info = pathinfo($attechment->file);
 			
-			
 			if($file_info['dirname'] != '.' && $file_info['dirname'] != '..')
-				$sub_dir = $file_info['dirname'];
+				$sub_dir = stripslashes_deep($file_info['dirname']);
 			
 			$uploads = wp_upload_dir(trim($sub_dir, '/')); // Array of key => value pairs	
 			$uploads_baseurl = $uploads['baseurl'];
@@ -1173,8 +1186,11 @@ function geodir_show_image( $request = array(), $size = 'thumbnail' ,$no_image =
 					$width_per = 100;
 			}
 			
+			//$html = '<div class="geodir_thumbnail" style="background-image:url(\''.$image->src.'\');"></div>';
+
 			//$html = '<div class="geodir_thumbnail"><img style="max-height:'. $max_size->h .'px;" alt="place image" src="' . $image->src . '"  /></div>';
-			if(is_admin()):
+			//print_r($_REQUEST);
+			if(is_admin() && !isset($_REQUEST['geodir_ajax'])):
 				$html = '<div class="geodir_thumbnail"><img style="max-height:'. $max_size->h .'px;" alt="place image" src="' . $image->src . '"  /></div>';
 			else : 
 				$html = '<div class="geodir_thumbnail" style="background-image:url(\''.$image->src.'\');"></div>';
@@ -1528,7 +1544,7 @@ function geodir_get_infowindow_html($postinfo_obj, $post_preview = ''){
 						$e++;
 						if($e==2){break;}// only show 3 event dates
 						$output .=  '<p>';
-						$geodir_num_dates++;
+						//$geodir_num_dates++;
 						if(isset($recuring_data['different_times']) && $recuring_data['different_times'] == '1'){
 							$starttimes = isset($recuring_data['starttimes'][$key]) ? $recuring_data['starttimes'][$key] : '';
 							$endtimes = isset($recuring_data['endtimes'][$key]) ? $recuring_data['endtimes'][$key] : '';
@@ -1556,27 +1572,30 @@ function geodir_get_infowindow_html($postinfo_obj, $post_preview = ''){
 					
 					echo $output;
 		}
-					?>
-             
-             
-             
-             
+		
+		
+					if($ID){
+		
+						$post_author = isset($postinfo_obj->post_author) ? $postinfo_obj->post_author : get_post_field( 'post_author', $ID );
+						?>
+
               <div class="geodir-bubble-meta-bottom">
                   <span class="geodir-bubble-rating"><?php echo $rating_star;?></span>
                   
-                  <span class="geodir-bubble-fav"><?php echo geodir_favourite_html($postinfo_obj->post_author,$ID);?></span>
+                  <span class="geodir-bubble-fav"><?php echo geodir_favourite_html($post_author,$ID);?></span>
                   <span class="geodir-bubble-reviews"><a href="<?php echo get_comments_link( $ID ); ?>" class="geodir-pcomments"><i class="fa fa-comments"></i>
                             <?php echo get_comments_number( $ID ); ?>
                                      </a></span>
               </div>
  
-              
+       <?php }?> 
+			       
 			</div>             					
 		</div>
 	</div>
 	<?php 
 	$html = ob_get_clean();
-	
+	$html = apply_filters('geodir_custom_infowindow_html' ,$html ,$postinfo_obj, $post_preview   ) ;
 	return $html;
 	}
 }
@@ -1693,6 +1712,11 @@ function geodir_delete_listing_info($deleted_postid, $force = false){
 	global $wpdb,$plugin_prefix;
 	
 	$post_type = get_post_type( $deleted_postid );
+	
+	$all_postypes = geodir_get_posttypes();
+
+	if(!in_array($post_type, $all_postypes))
+		return false;
 	
 	$table = $plugin_prefix . $post_type . '_detail';
 	
@@ -2051,8 +2075,10 @@ function geodir_set_wp_featured_image($post_id)
 			$id = wp_insert_attachment( $attachment, $filename, $post_id );
 			
 			if ( ! is_wp_error( $id ) ) {
-				
+			
 				set_post_thumbnail($post_id,$id);
+				
+				require_once(ABSPATH . 'wp-admin/includes/image.php');
 				wp_update_attachment_metadata( $id, wp_generate_attachment_metadata( $id, $filename ) );
 				
 			}
@@ -2133,7 +2159,49 @@ add_action('wp_ajax_gd_copy_original_translation', 'gd_copy_original_translation
 
 
 
-
+function geodir_get_custom_fields_type($listing_type = ''){
+	
+	global $wpdb;
+	
+	if($listing_type == '')
+		$listing_type = 'gd_place';
+	
+	$fields_info = array();
+	
+	$get_data = $wpdb->get_results(
+		$wpdb->prepare(
+			"SELECT htmlvar_name, field_type, extra_fields FROM ".GEODIR_CUSTOM_FIELDS_TABLE." WHERE post_type=%s AND is_active='1'",
+			array($listing_type)
+		)
+	);
+	
+	if(!empty($get_data)){
+		
+		foreach($get_data as $data){
+			
+			if($data->field_type == 'address'){
+				
+				$extra_fields = unserialize($data->extra_fields);
+				
+				$prefix = $data->htmlvar_name.'_';
+				
+				$fields_info[$prefix.'address'] = $data->field_type;
+				
+				if(isset($extra_fields['show_zip']) && $extra_fields['show_zip'])
+					$fields_info[$prefix.'zip'] = $data->field_type;
+					
+			}else{
+				
+				$fields_info[$data->htmlvar_name] = $data->field_type;
+				
+			}
+			
+		}
+		
+	}
+	
+	return apply_filters('geodir_get_custom_fields_type', $fields_info, $listing_type);
+}
